benefitrates implemented

+++ app/products/admin/resources/CreateTables.py
@@ -1,3 +1,4 @@
+from flask import request
 from flask_restful import Resource
 from app.extensions import db
 
@@ -7,7 +8,9 @@ class CreateTables(Resource):
     @classmethod
     def get(cls):
         try:
-            db.drop_all()
+            dropTables = request.args.get('drop')
+            if dropTables == 'Y':
+                db.drop_all()
             db.create_all()
         except Exception as e:
             return str(e), 400
